Move JsonFields to readers.schema
